Move tab-bar to components
