+++ src/settings/solid-js/components/m-tab-bar.tsx
